requirements.txt
fastapi
uvicorn
mistralai
weave
pydantic
python-dotenv
requests
gTTS
